refactor: better naming and remove unused module
