+++ src/commands/cleanup.py
