+++ src/commands/encrypt.py
