+++ src/commands/extract.py
