+++ src/commands/info.py
